Epic_Cross_game/requirements.txt
#The requirements.txt file is used to document and manage the external Python libraries and their versions.
It ensures consistency, reproducibility, and compatibility of dependencies across different environments.
It also makes it easier to install and maintain the required libraries for a project.



bcrypt==4.0.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Columnar==1.4.1
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
Flask==3.0.0
fonttools==4.46.0
hupper==1.12
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mysql-connector-python==8.1.0
numpy==1.26.2
packaging==23.2
passlib==1.7.4
PasteDeploy==3.1.0
Pillow==10.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
player==0.6.1
protobuf==4.21.12
pygame==2.5.2
pyparsing==3.1.1
pyramid==2.0.2
pyramid-jinja2==2.10
python-dateutil==2.8.2
requests==2.31.0
self==2020.12.3
six==1.16.0
toolz==0.12.0
translationstring==1.4
urllib3==2.0.7
venusian==3.1.0
wcwidth==0.2.8
WebOb==1.8.7
Werkzeug==3.0.0
zope.deprecation==5.0
zope.interface==6.1
